+++ pages/_staticPage.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import events from '~/components/events/events'
+import events from '~/lib/events'
 import loadData from '~/lib/load-data'
 
 export default {
