+++ pages/index.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import events from '~/components/events/events'
+import events from '~/lib/events'
 import loadData from '~/lib/load-data'
 import ThemeList from '~/components/theme-list/ThemeList'
